% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{wkverify_command}
\alias{wkverify_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkverify_command(verifier, cmd_json)
}
\arguments{
\item{verifier}{A verifier handle from [wkverify_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
